refactor : 기타수정
