+++ src/main/java/kr/hhplus/ecommerce/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package kr.hhplus.ecommerce.config.swagger;
+package kr.hhplus.ecommerce.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
