+++ src/main/resources/data.sql
@@ -33,22 +33,22 @@ VALUES
     (1, 1, 100000, NOW(), NOW(), 0);
 
 -- 🔥 포인트 이력 (point_history) 데이터
-INSERT INTO point_history (point_history_id, point_id, amount, transaction_type, registered_at, updated_at)
-VALUES
-    (1, 1, 50000, 'CHARGE', NOW(), NOW());
-
--- 🔥 주문 (order) 데이터
-INSERT INTO `order` (id, user_id, total_amount, discount_amount, payment_amount, registered_at, updated_at, status, issued_coupon_id)
-VALUES
-    (1, 1, 1200000, 5000, 1150000, NOW(), NOW(), 'CREATED', 1);
-
--- 🔥 주문 상품 (order_item) 데이터
-INSERT INTO order_item (id, order_id, product_option_id, quantity, unit_price, registered_at, updated_at, status)
-VALUES
-    (1, 1, 101, 1, 1000000, NOW(), NOW(), 'CREATED'),
-    (2, 1, 102, 2, 1200000, NOW(), NOW(), 'CREATED');
-
--- 🔥 결제 (payment) 데이터
-INSERT INTO payment (id, order_id, amount, registered_at, updated_at, paid_at, status)
-VALUES
-    (1, 1, 1150000, NOW(), NOW(), NOW(), 'WAITING');
+# INSERT INTO point_history (point_history_id, point_id, amount, transaction_type, registered_at, updated_at)
+# VALUES
+#     (1, 1, 50000, 'CHARGE', NOW(), NOW());
+#
+# -- 🔥 주문 (order) 데이터
+# INSERT INTO `order` (id, user_id, total_amount, discount_amount, payment_amount, registered_at, updated_at, status, issued_coupon_id)
+# VALUES
+#     (1, 1, 1200000, 5000, 1150000, NOW(), NOW(), 'CREATED', 1);
+#
+# -- 🔥 주문 상품 (order_item) 데이터
+# INSERT INTO order_item (id, order_id, product_option_id, quantity, unit_price, registered_at, updated_at, status)
+# VALUES
+#     (1, 1, 101, 1, 1000000, NOW(), NOW(), 'CREATED'),
+#     (2, 1, 102, 2, 1200000, NOW(), NOW(), 'CREATED');
+#
+# — 🔥 결제 (payment) 데이터
+# INSERT INTO payment (id, order_id, amount, registered_at, updated_at, paid_at, status)
+# VALUES
+#     (1, 1, 1150000, NOW(), NOW(), NOW(), 'WAITING');
